java homework 2 making some changes
